correlation count

+++ cpp/src/dmrpt/algo/drpt_global.cpp
@@ -512,7 +512,7 @@ dmrpt::DRPTGlobal::calculate_tree_leaf_correlation() {
                 int size = data_points.size();
                 std::transform(correlation_matrix[tree][leaf][c].begin(), correlation_matrix[tree][leaf][c].end(),
                                correlation_matrix[tree][leaf][c].begin(), [&](float x){(x/size)*100;});
-                int* max_element = std::max_element(correlation_matrix[tree][leaf][c].begin(),
+                float* max_element = std::max_element(correlation_matrix[tree][leaf][c].begin(),
                                                      correlation_matrix[tree][leaf][c].end());
                 int selected_leaf = max_element - correlation_matrix[tree][leaf][c].begin();
